Remove duplicate module
