--- coordinator/src/services/signature-verification.ts
@@ -1,96 +1,0 @@
-import {ElementCompact} from "xml-js"
-import {hl7V3} from "@models"
-import {writeXmlStringCanonicalized} from "./serialisation/xml"
-import {convertFragmentsToHashableFormat, extractFragments} from "./translation/request/signature"
-import {createParametersDigest} from "./translation/request"
-import crypto from "crypto"
-import {isTruthy} from "./translation/common"
-
-export function verifySignature(parentPrescription: hl7V3.ParentPrescription): Array<string> {
-  const validSignatureFormat = verifySignatureHasCorrectFormat(parentPrescription)
-  if (!validSignatureFormat) {
-    return ["Invalid signature format."]
-  }
-
-  const errors = []
-
-  const validSignature = verifyPrescriptionSignatureValid(parentPrescription)
-  if (!validSignature) {
-    errors.push("Signature is invalid.")
-  }
-
-  const matchingSignature = verifySignatureDigestMatchesPrescription(parentPrescription)
-  if (!matchingSignature) {
-    errors.push("Signature doesn't match prescription.")
-  }
-
-  const cerificateIsValid = verifyCertificate(parentPrescription)
-  if (!cerificateIsValid) {
-    errors.push("Certificate is invalid.")
-  }
-
-  return errors
-}
-
-export function verifySignatureHasCorrectFormat(parentPrescription: hl7V3.ParentPrescription): boolean {
-  const signatureRoot = extractSignatureRootFromParentPrescription(parentPrescription)
-  const signature = signatureRoot?.Signature
-  const signedInfo = signature?.SignedInfo
-  const signatureValue = signature?.SignatureValue?._text
-  const x509Certificate = signature?.KeyInfo?.X509Data?.X509Certificate?._text
-  return isTruthy(signedInfo) && isTruthy(signatureValue) && isTruthy(x509Certificate)
-}
-
-export function verifySignatureDigestMatchesPrescription(parentPrescription: hl7V3.ParentPrescription): boolean {
-  const signatureRoot = extractSignatureRootFromParentPrescription(parentPrescription)
-  const digestOnPrescription = extractDigestFromSignatureRoot(signatureRoot)
-  const calculatedDigestFromPrescription = calculateDigestFromParentPrescription(parentPrescription)
-  console.log(`Digest on Prescription: ${digestOnPrescription}`)
-  console.log(`Calculated digest from Prescription: ${calculatedDigestFromPrescription}`)
-  return digestOnPrescription === calculatedDigestFromPrescription
-}
-
-export function verifyPrescriptionSignatureValid(parentPrescription: hl7V3.ParentPrescription): boolean {
-  const signatureRoot = extractSignatureRootFromParentPrescription(parentPrescription)
-  return verifySignatureValid(signatureRoot)
-}
-
-export function extractSignatureRootFromParentPrescription(
-  parentPrescription: hl7V3.ParentPrescription
-): ElementCompact {
-  const pertinentPrescription = parentPrescription.pertinentInformation1.pertinentPrescription
-  return pertinentPrescription.author.signatureText
-}
-
-function extractDigestFromSignatureRoot(signatureRoot: ElementCompact) {
-  const signature = signatureRoot.Signature
-  const signedInfo = signature.SignedInfo
-  signedInfo._attributes = {
-    xmlns: signature._attributes.xmlns
-  }
-  return writeXmlStringCanonicalized({SignedInfo: signedInfo})
-}
-
-function calculateDigestFromParentPrescription(parentPrescription: hl7V3.ParentPrescription) {
-  const fragments = extractFragments(parentPrescription)
-  const fragmentsToBeHashed = convertFragmentsToHashableFormat(fragments)
-  const digestFromPrescriptionBase64 = createParametersDigest(fragmentsToBeHashed)
-  return Buffer.from(digestFromPrescriptionBase64, "base64").toString("utf-8")
-}
-
-function verifySignatureValid(signatureRoot: ElementCompact) {
-  const signatureVerifier = crypto.createVerify("RSA-SHA1")
-  const digest = extractDigestFromSignatureRoot(signatureRoot)
-  signatureVerifier.update(digest)
-  const signature = signatureRoot.Signature
-  const signatureValue = signature.SignatureValue._text
-  const x509Certificate = signature.KeyInfo.X509Data.X509Certificate._text
-  const x509CertificatePem = `-----BEGIN CERTIFICATE-----\n${x509Certificate}\n-----END CERTIFICATE-----`
-  return signatureVerifier.verify(x509CertificatePem, signatureValue, "base64")
-}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function verifyCertificate(parentPrescription: hl7V3.ParentPrescription) {
-  console.log("Skipping certificate verification...")
-  return true
-}
