+++ coordinator/src/services/verification/signature-verification.ts
@@ -11,15 +11,24 @@ export function verifySignature(parentPrescription: hl7V3.ParentPrescription): A
   if (!validSignatureFormat) {
     return ["Invalid signature format."]
   }
+
   const errors = []
+
   const validSignature = verifyPrescriptionSignatureValid(parentPrescription)
   if (!validSignature) {
     errors.push("Signature is invalid.")
   }
+
   const matchingSignature = verifySignatureDigestMatchesPrescription(parentPrescription)
   if (!matchingSignature) {
     errors.push("Signature doesn't match prescription.")
   }
+
+  const cerificateIsValid = verifyCertificate(parentPrescription)
+  if (!cerificateIsValid) {
+    errors.push("Certificate is invalid.")
+  }
+
   return errors
 }
 
@@ -78,4 +87,10 @@ function verifySignatureValid(signatureRoot: ElementCompact) {
   const x509Certificate = signature.KeyInfo.X509Data.X509Certificate._text
   const x509CertificatePem = `-----BEGIN CERTIFICATE-----\n${x509Certificate}\n-----END CERTIFICATE-----`
   return signatureVerifier.verify(x509CertificatePem, signatureValue, "base64")
+}
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+function verifyCertificate(parentPrescription: hl7V3.ParentPrescription) {
+  console.log("Skipping certificate verification...")
+  return true
 }
